Separate map from LocationsPane to new OverviewPane
