+++ src/components/sections/maps/MapOverviewPane.scss
@@ -1,0 +1,9 @@
+.MapOverviewPane {
+    .PaneBase-header {
+        display: none;
+    }
+
+    .PaneBase-content {
+        margin: 0 !important;
+    }
+}
